♻️ Refactor: #19 사이드바 모듈화
