+++ play.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>PlayPage</title>
     <link rel="stylesheet" href="./src/styles/play.css" />
+    <link rel="stylesheet" href="./src/styles/sidebar.css" />
     <link rel="stylesheet" href="./src/styles/common.css" />
     <link
       rel="stylesheet"
@@ -12,7 +13,7 @@
     />
   </head>
   <body>
-    <main class="play-wrap">
+    <main class="main-container play-wrap">
       <section class="play-cont">
         <div id="foldFailToastMessage"></div>
         <canvas class="play-canvas"></canvas>
@@ -82,11 +83,11 @@
             <button class="round-button button-purple restart-button">
               RESTART
             </button>
-            <button class="round-button button-red">SHARE</button>
+            <button class="round-button button-red share-button">SHARE</button>
           </div>
         </div>
       </section>
-      <nav class="play-nav">
+      <nav class="side-nav">
         <div>
           <button class="home-button nav-button"></button>
           <button class="mode-button nav-button"></button>
@@ -166,6 +167,7 @@
       </aside>
     </main>
     <script type="module" src="/src/components/play.js" defer></script>
+    <script type="module" src="/src/components/sidebar.js" defer></script>
     <script
       type="module"
       src="src/components/modules/origami.js"
