+++ src/styles/common.css
@@ -12,6 +12,38 @@ body {
   background-color: #343394;
 }
 
+.main-container {
+  display: flex;
+  align-items: center;
+  width: 100vw;
+  height: 100vh;
+  background-color: #343394;
+}
+
+.main-container section {
+  position: relative;
+  display: flex;
+  align-items: flex-end;
+  width: 100%;
+  height: 96%;
+  margin-left: 20px;
+  border-radius: 30px;
+  background-color: #e7e6ff;
+}
+
+section.active:after {
+  position: absolute;
+  z-index: 2;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  border-radius: 30px;
+  background-color: #fff;
+  opacity: 0.7;
+  content: '';
+}
+
 button {
   cursor: pointer;
   border: none;
