+++ src/styles/sidebar.css
@@ -1,0 +1,225 @@
+.side-nav {
+  position: relative;
+  z-index: 3;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  width: 110px;
+  height: 100%;
+  box-sizing: border-box;
+  padding: 60px 0;
+  background-color: #343394;
+}
+
+.side-nav div button {
+  display: block;
+  margin-bottom: 15px;
+  text-align: center;
+}
+
+.side-nav div button:last-child {
+  margin-bottom: 0;
+}
+
+.info-wrap {
+  position: absolute;
+  z-index: 3;
+  bottom: 160px;
+  right: 120px;
+  display: none;
+  border-radius: 30px;
+  background-color: #fff;
+  transition: all 0.2s ease-in-out;
+}
+
+.info-wrap::before {
+  position: absolute;
+  z-index: 3;
+  bottom: -54px;
+  right: -32px;
+  content: '';
+  width: 61px;
+  height: 79px;
+  background: url('/src/assets/img/speech-bubble.png') no-repeat -5px -30px;
+  background-size: contain;
+}
+
+.info-wrap.visible {
+  display: block;
+}
+
+.side-nav .info-wrap p {
+  width: 745px;
+  padding: 30px;
+  line-height: 1.8;
+  text-align: left;
+}
+
+.side-nav .info-wrap p span {
+  color: #343394;
+  font-weight: 900;
+}
+
+.home-button {
+  width: 60px;
+  height: 60px;
+  background: url('/src/assets/img/nav_sprites.png') no-repeat -5px -8px;
+  opacity: 0.5;
+}
+
+.mode-button {
+  width: 60px;
+  height: 60px;
+  background: url('/src/assets/img/nav_sprites.png') no-repeat -184px -80px;
+  opacity: 0.5;
+}
+
+.mode-button.active {
+  background: #fff url('/src/assets/img/checked_sprites.png') no-repeat -8px -3px;
+  opacity: 1;
+}
+
+.mode-button.active:hover {
+  background-color: #fff;
+  opacity: 1;
+}
+
+.gallery-button {
+  width: 60px;
+  height: 60px;
+  background: url('/src/assets/img/nav_sprites.png') no-repeat -97px -8px;
+  opacity: 0.5;
+}
+
+.sound-button {
+  width: 60px;
+  height: 60px;
+  background: url('/src/assets/img/nav_sprites.png') no-repeat -100px -80px;
+  opacity: 0.5;
+  transition: none;
+}
+
+.info-button {
+  width: 60px;
+  height: 60px;
+  margin-bottom: 0;
+  background: url('/src/assets/img/nav_sprites.png') no-repeat -4px -81px;
+  opacity: 0.5;
+}
+
+.info-button.active {
+  background: url('/src/assets/img/checked_sprites.png') no-repeat -52px -55px;
+  background-color: #fff;
+  opacity: 1;
+}
+
+.info-button.active:hover {
+  background-color: #fff;
+  opacity: 1;
+}
+
+.nav-button {
+  border-radius: 999px;
+  transition: none;
+}
+
+.nav-button:hover {
+  width: 60px;
+  height: 60px;
+  padding: 10px;
+  border-radius: 999px;
+  background-color: #4746c1;
+  opacity: 80%;
+}
+
+.small-start-button {
+  width: 210px;
+  height: 60px;
+  border-radius: 10px;
+}
+
+.mode-sidebar {
+  position: absolute;
+  z-index: 1;
+  top: 2%;
+  right: -250px;
+  height: 96%;
+  overflow: hidden;
+  box-sizing: border-box;
+  width: 0;
+  padding: 30px 0 0 0;
+  border-radius: 30px;
+  background-color: #fff;
+  opacity: 0;
+  filter: drop-shadow(0 1px 1px rgba(0, 0, 0, 0.25));
+  transition: all 0.2s ease-in-out;
+}
+
+.mode-sidebar.visible {
+  right: 102px;
+  width: 269px;
+  opacity: 1;
+}
+
+.mode-sidebar ul li {
+  margin-bottom: 35px;
+  padding: 0 30px;
+  transition: all 0.2s ease-in-out;
+}
+
+.mode-sidebar ul li.active {
+  padding: 30px 30px;
+  background-color: rgba(52, 51, 148, 0.1);
+}
+
+.mode-sidebar ul li:hover {
+  padding: 30px 30px;
+  background-color: rgba(52, 51, 148, 0.1);
+}
+
+.mode-sidebar ul li div {
+  position: relative;
+  width: 209px;
+  height: 155px;
+  border-radius: 30px;
+  background-image: url(../assets/img/green-gamebg.png);
+  background-size: contain;
+}
+
+.mode-sidebar ul li h3 {
+  position: absolute;
+  bottom: 30px;
+  left: 20px;
+  color: #fff;
+  font-size: 24px;
+  font-weight: 900;
+}
+
+.mode-sidebar ul li p {
+  position: absolute;
+  z-index: 1;
+  top: -20px;
+  right: -15px;
+  width: 120px;
+  height: 120px;
+}
+
+.mode-sidebar ul li p img {
+  width: 100%;
+  height: 100%;
+}
+
+.mode-sidebar ul li:last-child {
+  margin-bottom: 0;
+}
+
+.mode-sidebar .sidebar-footer {
+  position: absolute;
+  z-index: 1;
+  bottom: 0;
+  width: 100%;
+  padding: 45px 0;
+  background-color: #fff;
+  text-align: center;
+}
